Trying to migrate raytrace to ComputeRHS and finalize
